+++ src/PrettyJson.js
@@ -17,15 +17,27 @@ export default class PrettyJson extends React.PureComponent {
         // @since 1.0.3 legacy alias for "data"
         json: PropTypes.oneOfType([PropTypes.object, PropTypes.string]),
         onPrettyPrint: PropTypes.func,
-        onError: PropTypes.func,
+        onError: PropTypes.func
     };
     static defaultProps = {
-        json: {},
+        json: {}
     };
+    warnLegacyProp = process.env.NODE_ENV !== 'production'
+        ? (oldProp, newProp) => console.warn(`[PrettyJson] The prop "${oldProp}" is deprecated. Please use "${newProp}" instead.`)
+        : null;
+    componentDidMount() {
+        if (process.env.NODE_ENV !== 'production') {
+            if (this.props.json) {
+                this.warnLegacyProp('json', 'data');
+            }
+        }
+    }
     componentDidUpdate(prevProps) {
         if (prevProps.json !== this.props.json || prevProps.data !== this.props.data) {
-            if (this.props.json && process.env.NODE_ENV !== 'production') {
-                console.warn('[PrettyJson] The "json" prop is deprecated. Please use "data" instead.')
+            if (process.env.NODE_ENV !== 'production') {
+                if (this.props.json) {
+                    this.warnLegacyProp('json', 'data');
+                }
             }
             this.convertHtml(this.elementRef);
         }
@@ -57,7 +69,7 @@ export default class PrettyJson extends React.PureComponent {
         return r + (pEnd || '');
     }
     prettyPrint(obj) {
-        try {    
+        try {
             obj = typeof obj === 'string' ? JSON.parse(obj) : pojo(obj);
             var jsonLine = /^( *)("[\w]+": )?("[^"]*"|[\w.+-]*)?([,[{])?$/gm;
             const html = JSON.stringify(obj, null, 3)
@@ -67,11 +79,10 @@ export default class PrettyJson extends React.PureComponent {
                 .replace(/>/g, '&gt;')
                 .replace(jsonLine, this.replacer);
             if (this.props.onPrettyPrint) {
-                this.props.onPrettyPrint(html)
+                this.props.onPrettyPrint(html);
             }
             return html;
-        }
-        catch (error) {
+        } catch (error) {
             if (this.props.onError) {
                 this.props.onError(error);
             }