+++ src/index.js
@@ -1,3 +1,3 @@
 import PrettyJson from "./PrettyJson";
 export default PrettyJson;
-
+export {default as pojo} from './pojo';