+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.0.5
+
+* add warning to componentDidMount
+
+# 1.0.4
+
+* export 'pojo' helper
+
 # 1.0.3
 
 * add `data` prop, legacy support `json` prop